airlines/requirements.txt
asgiref==3.9.1
crispy-bootstrap4==2025.6
Django==5.2.5
django-crispy-forms==2.4
pillow==11.3.0
python-decouple==3.8
sqlparse==0.5.3

djangorestframework==3.15.2